Issue #14 Change @Order type to any Number type

+++ src/test/java/org/jboss/qa/phaser/phase/third/Third.java
@@ -20,5 +20,5 @@ public @interface Third {
 	String second() default "";
 
 	@Order
-	int order() default 0;
+	double order() default 0;
 }
